+++ DbScripts.md
@@ -1,9 +1,9 @@
 # DB Scripts
 
-* Create Statuses collection
+* Create Stages collection
 
 ```sh
-db.statuses.insert({ displayOrder : "1", key : "assigned", value: "Assinged"})
-db.statuses.insert({ displayOrder : "2", key : "inProgress", value: "In Progress"})
-db.statuses.insert({ displayOrder : "3", key : "closed", value: "Closed"})
+db.stages.insert({ displayOrder : "1", key : "assigned", value: "Assinged"})
+db.stages.insert({ displayOrder : "2", key : "inProgress", value: "In Progress"})
+db.stages.insert({ displayOrder : "3", key : "closed", value: "Closed"})
 ```
